implements method tiempoEstimadoParaAtencion

+++ src/main/java/com/jc/sgtasec/web/dto/RestAtencionDto.java
@@ -24,6 +24,6 @@ public class RestAtencionDto {
 	
 	private Long cantidadMaximaLlamadas;
 	
-	private int tiempoEstimadoParaAtencion;
+	private Long tiempoEstimadoParaAtencion;
 
 }
